feat: altered services to helper files, create new file validationForms
